Broken check in, incorrect positioning.
git-svn-id: http://svn.gareth.se/swm/trunk@433 add37ce8-0dd3-4abe-8994-b1afb2f1f0e5

+++ swm/src/se/gareth/swm/Animation.java
@@ -128,11 +128,11 @@ public class Animation {
 		return mRunning;
 	}
 	
-	public double getWidth() {
+	public int getWidth() {
 		return mSprite.getFrameWidth();
 	}
 	
-	public double getHeight() {
+	public int getHeight() {
 		return mSprite.getFrameHeight();
 	}
 }
